requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
folium
streamlit-folium
joblib
Flask
psycopg2-binary